+++ shamrock/__init__.py
@@ -5,5 +5,5 @@ from .shamrock import ENDPOINTS, NAVIGATION, Shamrock
 
 __author__ = "Zlatko Mašek"
 __email__ = "<zlatko.masek@gmail.com>"
-__version__ = "0.0.1"
+__version__ = "0.1.0"
 __all__ = ["Shamrock", "ENDPOINTS", "NAVIGATION"]